The decide_session function is now in views since that is the only place it is needed outside of tests.
