+++ tests/test_decide_session.py
@@ -1,6 +1,6 @@
 from django.test import TransactionTestCase  # , TestCase
 from django.db.models.query import QuerySet
-from classwork.forms import decide_session
+from classwork.views import decide_session
 from classwork.models import Session
 from datetime import date, timedelta
