+++ web/classwork/views.py
@@ -2,18 +2,49 @@ from django.views.generic import ListView, CreateView, DetailView, UpdateView
 # from django.contrib.auth.mixins import LoginRequiredMixin
 # from django.urls import reverse_lazy
 # from collections import OrderedDict
-from .forms import RegisterForm, PaymentForm, decide_session  # , ProfileForm, UserForm
 from django.shortcuts import get_object_or_404, redirect  # used for Payments
 from django.template.response import TemplateResponse  # used for Payments
 from payments import get_payment_model, RedirectNeeded  # used for Payments
 from django.db.models import Q
 from django.core.exceptions import ObjectDoesNotExist
 from django.contrib.auth import get_user_model
+from .forms import RegisterForm, PaymentForm  # , ProfileForm, UserForm
 from .models import (SiteContent, Resource, Location, ClassOffer, Subject,  # ? Session,
-                     Profile, Payment, Registration)
+                     Profile, Payment, Registration, Session)
+from datetime import datetime
 User = get_user_model()
 
 # TODO: Clean out excessive print lines telling us where we are.
+
+
+def decide_session(sess=None, display_date=None):
+    """ Typically we want to see the current session (returned if no params set)
+        Sometimes we want to see a future session.
+        Used by many views, generally those that need a list of ClassOffers
+        that a user can view, sign up for, get a check-in sheet, pay for, etc.
+    """
+    # print('======= forms function - decide_session ==========')
+    sess_data = []
+    if sess is None:
+        target = display_date or datetime.now()
+        # TODO: Fix for dealing with Sessions do not have an expire_date
+        sess_data = Session.objects.filter(publish_date__lte=target, expire_date__gte=target)
+    else:
+        if display_date:
+            raise SyntaxError("You can't filter by both Session and Date")
+        if sess == 'all':
+            return Session.objects.all()
+        if not isinstance(sess, list):
+            sess = [sess]
+        try:
+            sess_data = Session.objects.filter(name__in=sess)
+        except TypeError:
+            sess_data = []
+    if not sess_data:
+        # TODO: Filter out future but not yet published Sessions
+        sess_data = [Session.objects.order_by('-key_day_date').first()]
+    # print(sess_data)
+    return sess_data  # a list of Session records, even if only 0-1 session
 
 
 class AboutUsListView(ListView):
